added cas-api module to separate cas public keys in core
