+++ cas-server-core/src/test/java/org/jasig/cas/authentication/principal/ResponseTests.java
@@ -37,7 +37,7 @@ public class ResponseTests {
         final String url = "http://localhost:8080/foo";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "foobar");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals(url + "?ticket=foobar", response.getUrl());
     }
 
@@ -46,7 +46,7 @@ public class ResponseTests {
         final String url = "http://localhost:8080/foo?test=boo";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "foobar");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals(url + "&ticket=foobar", response.getUrl());
     }
 
@@ -55,7 +55,7 @@ public class ResponseTests {
         final String url = "http://localhost:8080/foo?test=boo#hello";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "foobar");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals("http://localhost:8080/foo?test=boo&ticket=foobar#hello", response.getUrl());
     }
 
@@ -64,7 +64,7 @@ public class ResponseTests {
         final String url = "http://localhost:8080/foo#hello";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "foobar");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals("http://localhost:8080/foo?ticket=foobar#hello", response.getUrl());
 
     }
@@ -74,7 +74,7 @@ public class ResponseTests {
         final String url = "https://www.example.com\r\nLocation: javascript:\r\n\r\n<script>alert(document.cookie)</script>";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "ST-12345");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals("https://www.example.com Location: javascript: <script>alert(document.cookie)</script>?ticket=ST-12345",
                 response.getUrl());
     }
@@ -84,7 +84,7 @@ public class ResponseTests {
         final String url = "https://www.example.com/πολιτικῶν";
         final Map<String, String> attributes = new HashMap<>();
         attributes.put("ticket", "ST-12345");
-        final Response response = Response.getRedirectResponse(url, attributes);
+        final DefaultResponse response = DefaultResponse.getRedirectResponse(url, attributes);
         assertEquals("https://www.example.com/πολιτικῶν?ticket=ST-12345", response.getUrl());
     }
 }
