+++ cas-server-core/src/test/java/org/jasig/cas/authentication/principal/SimpleWebApplicationServiceImplTests.java
@@ -20,7 +20,7 @@ package org.jasig.cas.authentication.principal;
 
 import static org.junit.Assert.*;
 
-import org.jasig.cas.authentication.principal.Response.ResponseType;
+import org.jasig.cas.authentication.principal.DefaultResponse.ResponseType;
 import org.junit.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 
@@ -39,7 +39,7 @@ public class SimpleWebApplicationServiceImplTests {
         request.setParameter("service", "service");
         final SimpleWebApplicationServiceImpl impl = SimpleWebApplicationServiceImpl.createServiceFrom(request);
 
-        final Response response = impl.getResponse("ticketId");
+        final DefaultResponse response = impl.getResponse("ticketId");
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
     }
@@ -67,7 +67,7 @@ public class SimpleWebApplicationServiceImplTests {
         request.setParameter("service", "service");
         final WebApplicationService impl = SimpleWebApplicationServiceImpl.createServiceFrom(request);
 
-        final Response response = impl.getResponse(null);
+        final DefaultResponse response = impl.getResponse(null);
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
         assertFalse(response.getUrl().contains("ticket="));
@@ -79,7 +79,7 @@ public class SimpleWebApplicationServiceImplTests {
         request.setParameter("service", "http://foo.com/");
         final WebApplicationService impl = SimpleWebApplicationServiceImpl.createServiceFrom(request);
 
-        final Response response = impl.getResponse(null);
+        final DefaultResponse response = impl.getResponse(null);
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
         assertFalse(response.getUrl().contains("ticket="));
@@ -92,7 +92,7 @@ public class SimpleWebApplicationServiceImplTests {
         request.setParameter("service", "http://foo.com/?param=test");
         final WebApplicationService impl = SimpleWebApplicationServiceImpl.createServiceFrom(request);
 
-        final Response response = impl.getResponse(null);
+        final DefaultResponse response = impl.getResponse(null);
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
         assertEquals("http://foo.com/?param=test", response.getUrl());
