+++ cas-server-support-saml/src/main/java/org/jasig/cas/support/saml/authentication/principal/GoogleAccountsService.java
@@ -20,7 +20,7 @@ package org.jasig.cas.support.saml.authentication.principal;
 
 
 import org.jasig.cas.authentication.principal.AbstractWebApplicationService;
-import org.jasig.cas.authentication.principal.Response;
+import org.jasig.cas.authentication.principal.DefaultResponse;
 import org.jasig.cas.services.RegisteredService;
 import org.jasig.cas.services.ServicesManager;
 import org.jasig.cas.support.saml.util.AbstractSaml20ObjectBuilder;
@@ -151,7 +151,7 @@ public class GoogleAccountsService extends AbstractWebApplicationService {
     }
 
     @Override
-    public Response getResponse(final String ticketId) {
+    public DefaultResponse getResponse(final String ticketId) {
         final Map<String, String> parameters = new HashMap<>();
         final String samlResponse = constructSamlResponse();
         final String signedResponse = BUILDER.signSamlResponse(samlResponse,
@@ -159,7 +159,7 @@ public class GoogleAccountsService extends AbstractWebApplicationService {
         parameters.put(SamlProtocolConstants.PARAMETER_SAML_RESPONSE, signedResponse);
         parameters.put(SamlProtocolConstants.PARAMETER_SAML_RELAY_STATE, this.relayState);
 
-        return Response.getPostResponse(getOriginalUrl(), parameters);
+        return DefaultResponse.getPostResponse(getOriginalUrl(), parameters);
     }
 
     /**
