+++ cas-server-support-saml/src/test/java/org/jasig/cas/support/saml/authentication/principal/GoogleAccountsServiceTests.java
@@ -19,7 +19,7 @@
 package org.jasig.cas.support.saml.authentication.principal;
 
 import org.jasig.cas.TestUtils;
-import org.jasig.cas.authentication.principal.Response;
+import org.jasig.cas.authentication.principal.DefaultResponse;
 import org.jasig.cas.authentication.principal.Service;
 import org.jasig.cas.services.DefaultRegisteredServiceUsernameProvider;
 import org.jasig.cas.services.RegisteredService;
@@ -97,8 +97,8 @@ public class GoogleAccountsServiceTests extends AbstractOpenSamlTests {
 
     @Test
     public void verifyResponse() {
-        final Response resp = this.googleAccountsService.getResponse("ticketId");
-        assertEquals(resp.getResponseType(), Response.ResponseType.POST);
+        final DefaultResponse resp = this.googleAccountsService.getResponse("ticketId");
+        assertEquals(resp.getResponseType(), DefaultResponse.ResponseType.POST);
         assertTrue(resp.getAttributes().containsKey(SamlProtocolConstants.PARAMETER_SAML_RESPONSE));
         assertTrue(resp.getAttributes().containsKey(SamlProtocolConstants.PARAMETER_SAML_RELAY_STATE));
     }
