+++ cas-server-support-saml/src/test/java/org/jasig/cas/support/saml/authentication/principal/SamlServiceTests.java
@@ -18,8 +18,8 @@
  */
 package org.jasig.cas.support.saml.authentication.principal;
 
-import org.jasig.cas.authentication.principal.Response;
-import org.jasig.cas.authentication.principal.Response.ResponseType;
+import org.jasig.cas.authentication.principal.DefaultResponse;
+import org.jasig.cas.authentication.principal.DefaultResponse.ResponseType;
 import org.jasig.cas.authentication.principal.WebApplicationService;
 import org.jasig.cas.support.saml.AbstractOpenSamlTests;
 import org.jasig.cas.support.saml.web.support.SamlArgumentExtractor;
@@ -41,7 +41,7 @@ public class SamlServiceTests extends AbstractOpenSamlTests {
         request.setParameter("TARGET", "service");
         final SamlService impl = SamlService.createServiceFrom(request);
 
-        final Response response = impl.getResponse("ticketId");
+        final DefaultResponse response = impl.getResponse("ticketId");
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
         assertTrue(response.getUrl().contains("SAMLart="));
@@ -62,7 +62,7 @@ public class SamlServiceTests extends AbstractOpenSamlTests {
         request.setParameter("TARGET", "service");
         final SamlService impl = SamlService.createServiceFrom(request);
 
-        final Response response = impl.getResponse(null);
+        final DefaultResponse response = impl.getResponse(null);
         assertNotNull(response);
         assertEquals(ResponseType.REDIRECT, response.getResponseType());
         assertFalse(response.getUrl().contains("SAMLart="));
